Cobbled together hacky search-by-album dialogue and download.
Works in general for popular artists with popular albums, however due to the heuristics used
in searching for the audio source using the metadata retrieved from Spotify (there is no way
I know of to match them one-to-one) it will ocassionally mismatch audio to metadata. Use the
search-by-song mode for finer-grained control.